fix deck-layers build
