+++ libs/deck-layers/src/index.ts
@@ -6,3 +6,4 @@ export * from './layers/context/ContextLayer'
 export * from './layers/context/context.hooks'
 export * from './layers/fourwings/FourwingsLayer'
 export * from './utils/layers'
+export * from './layer-composer/layer-composer.hooks'
