+++ package.json
@@ -33,10 +33,18 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@deck.gl/core": "^8.9.16",
+    "@deck.gl/extensions": "^8.9.16",
+    "@deck.gl/geo-layers": "^8.9.16",
+    "@deck.gl/layers": "^8.9.16",
+    "@deck.gl/mesh-layers": "^8.9.16",
+    "@deck.gl/react": "^8.9.16",
     "@globalfishingwatch/map-convert": "github:GlobalFishingWatch/map-convert",
     "@mapbox/tile-cover": "^3.0.2",
     "@mapbox/tilebelt": "^1.0.2",
     "@mapbox/vector-tile": "^1.3.1",
+    "@nebula.gl/edit-modes": "^1.0.4",
+    "@nebula.gl/layers": "^1.0.4",
     "@popperjs/core": "2.10.2",
     "@reduxjs/toolkit": "^1.9.3",
     "@tippyjs/react": "4.2.6",
@@ -68,7 +76,7 @@
     "mapbox-gl": "npm:empty-npm-package@1.0.0",
     "match-sorter": "^6.3.1",
     "memoize-one": "^6.0.0",
-    "next": "13.4.4-canary.10",
+    "next": "13.4.4",
     "node-fetch": "2.6.7",
     "papaparse": "5.3.2",
     "pbf": "^3.2.1",
@@ -194,13 +202,7 @@
     "node-fetch": "2.6.7",
     "underscore": "^1.12.1",
     "tslib": "^2.5.0",
-    "webpack-sources": "3.2.1",
-    "@deck.gl/core": "^8.8.24",
-    "@deck.gl/extensions": "^8.8.24",
-    "@deck.gl/geo-layers": "^8.8.24",
-    "@deck.gl/layers": "^8.8.24",
-    "@deck.gl/mesh-layers": "^8.8.24",
-    "@deck.gl/react": "^8.8.24"
+    "webpack-sources": "3.2.1"
   },
   "husky": {
     "hooks": {
